+++ src/MainService.cs
@@ -54,6 +54,12 @@ public class MainService
         Option<DateTime> lastModifyBeforeOption = CreateLastModifyBeforeOption();
         cmd.AddOption(lastModifyBeforeOption);
 
+        Option<DateTime> createAfterOption = CreateCreateAfterOption();
+        cmd.AddOption(createAfterOption);
+
+        Option<DateTime> createBeforeOption = CreateCreateBeforeOption();
+        cmd.AddOption(createBeforeOption);
+
         Argument<string> dirArg = new("dir", "The target directory to analysis.")
         {
             Arity = ArgumentArity.ExactlyOne
@@ -69,6 +75,8 @@ public class MainService
             bool enableVerbose = context.ParseResult.GetValueForOption(verboseOption);
             DateTime lastModifyAfter = context.ParseResult.GetValueForOption(lastModifyAfterOption);
             DateTime lastModifyBefore = context.ParseResult.GetValueForOption(lastModifyBeforeOption);
+            DateTime createAfter = context.ParseResult.GetValueForOption(createAfterOption);
+            DateTime createBefore = context.ParseResult.GetValueForOption(createBeforeOption);
 
             if (string.IsNullOrEmpty(dir) || !Directory.Exists(dir))
             {
@@ -76,11 +84,17 @@ public class MainService
                 return;
             }
 
-            DffOption option = new(output, includes, excludes, enableVerbose, _args)
+            DffOption option = new(_args)
             {
                 Dir = Path.GetFullPath(dir),
                 LastModifyAfter = lastModifyAfter,
-                LastModifyBefore = lastModifyBefore
+                LastModifyBefore = lastModifyBefore,
+                CreateAfter = createAfter,
+                CreateBefore = createBefore,
+                EnableVerbose = enableVerbose,
+                OutputDir = output,
+                Includes = includes,
+                Excludes = excludes
             };
 
             await using DffHandler handler = new(option, context.GetCancellationToken());
@@ -153,6 +167,12 @@ public class MainService
         };
         cmd.AddArgument(destArg);
 
+        Option<DateTime> createAfterOption = CreateCreateAfterOption();
+        cmd.AddOption(createAfterOption);
+
+        Option<DateTime> createBeforeOption = CreateCreateBeforeOption();
+        cmd.AddOption(createBeforeOption);
+
         cmd.SetHandler(async (context) =>
             {
                 string output = context.ParseResult.GetValueForOption(outputOption);
@@ -167,6 +187,8 @@ public class MainService
                 string fromFile = context.ParseResult.GetValueForOption(fromFileOption);
                 DateTime lastModifyAfter = context.ParseResult.GetValueForOption(lastModifyAfterOption);
                 DateTime lastModifyBefore = context.ParseResult.GetValueForOption(lastModifyBeforeOption);
+                DateTime createAfter = context.ParseResult.GetValueForOption(createAfterOption);
+                DateTime createBefore = context.ParseResult.GetValueForOption(createBeforeOption);
 
                 if (string.IsNullOrEmpty(output))
                 {
@@ -175,7 +197,7 @@ public class MainService
 
                 _ = Directory.CreateDirectory(output);
 
-                RmOption option = new(output, includes, excludes, enableVerbose, _args)
+                RmOption option = new(_args)
                 {
                     Destination = Path.GetFullPath(dest),
                     File = file,
@@ -184,7 +206,13 @@ public class MainService
                     Yes = yes,
                     FromFile = fromFile,
                     LastModifyAfter = lastModifyAfter,
-                    LastModifyBefore = lastModifyBefore
+                    LastModifyBefore = lastModifyBefore,
+                    CreateAfter = createAfter,
+                    CreateBefore = createBefore,
+                    EnableVerbose = enableVerbose,
+                    OutputDir = output,
+                    Includes = includes,
+                    Excludes = excludes
                 };
 
                 await using RmHandler handler = new(option, context.GetCancellationToken());
@@ -250,6 +278,24 @@ public class MainService
         };
         cmd.AddArgument(destDirArg);
 
+        Option<DateTime> createAfterOption = CreateCreateAfterOption();
+        cmd.AddOption(createAfterOption);
+
+        Option<DateTime> createBeforeOption = CreateCreateBeforeOption();
+        cmd.AddOption(createBeforeOption);
+
+        Option<bool> preserveCreateOption = new("--preserve-create", "Preserve creation time for destination file.")
+        {
+            Arity = ArgumentArity.ZeroOrOne
+        };
+        cmd.AddOption(preserveCreateOption);
+
+        Option<bool> preserveLastModifyOption = new("--preserve-last-modify", "Preserve last modify time for destination file.")
+        {
+            Arity = ArgumentArity.ZeroOrOne
+        };
+        cmd.AddOption(preserveLastModifyOption);
+
         cmd.SetHandler(async (context) =>
             {
                 string output = context.ParseResult.GetValueForOption(outputOption);
@@ -263,6 +309,10 @@ public class MainService
                 bool sizeOnly = context.ParseResult.GetValueForOption(sizeOnlyOption);
                 DateTime lastModifyAfter = context.ParseResult.GetValueForOption(lastModifyAfterOption);
                 DateTime lastModifyBefore = context.ParseResult.GetValueForOption(lastModifyBeforeOption);
+                DateTime createAfter = context.ParseResult.GetValueForOption(createAfterOption);
+                DateTime createBefore = context.ParseResult.GetValueForOption(createBeforeOption);
+                bool preserveCreate = context.ParseResult.GetValueForOption(preserveCreateOption);
+                bool preserveLastModify = context.ParseResult.GetValueForOption(preserveLastModifyOption);
 
                 if (string.IsNullOrEmpty(output))
                 {
@@ -271,15 +321,23 @@ public class MainService
 
                 _ = Directory.CreateDirectory(output);
 
-                SyncOption option = new(output, includes, excludes, enableVerbose, _args)
+                SyncOption option = new(_args)
                 {
                     Delete = delete,
                     DryRun = dryRun,
                     SrcDir = srcDir,
                     DestDir = destDir,
                     SizeOnly = sizeOnly,
+                    PreserveCreateTime = preserveCreate,
+                    PreserveLastModifyTime = preserveLastModify,
+                    LastModifyAfter = lastModifyAfter,
                     LastModifyBefore = lastModifyBefore,
-                    LastModifyAfter = lastModifyAfter
+                    CreateAfter = createAfter,
+                    CreateBefore = createBefore,
+                    EnableVerbose = enableVerbose,
+                    OutputDir = output,
+                    Includes = includes,
+                    Excludes = excludes
                 };
 
                 await using SyncHandler handler = new(option, context.GetCancellationToken());
@@ -291,12 +349,14 @@ public class MainService
 
     private Option<string> CreateOutputOption()
     {
-        return new Option<string>(new[] { "-o", "--output" },
+        Option<string> option = new(new[] { "-o", "--output" },
             "The output directory for logs or any file generated during processing.")
         {
             IsRequired = false,
             Arity = ArgumentArity.ZeroOrOne
         };
+
+        return option;
     }
 
     private Option<string[]> CreateIncludeOption()
@@ -328,7 +388,7 @@ public class MainService
 
     private Option<DateTime> CreateLastModifyAfterOption()
     {
-        return new Option<DateTime>(new[] { "--last-modify-after" }, "Last modify after filter. e.g., 2022-10-01T10:20:21")
+        return new Option<DateTime>(new[] { "--last-modify-after" }, "Last modify time after filter. e.g., 2022-10-01T10:20:21")
         {
             IsRequired = false,
             Arity = ArgumentArity.ZeroOrOne
@@ -337,7 +397,25 @@ public class MainService
 
     private Option<DateTime> CreateLastModifyBeforeOption()
     {
-        return new Option<DateTime>(new[] { "--last-modify-before" }, "Last modify before fitler. e.g., 2022-08-02T16:20:21")
+        return new Option<DateTime>(new[] { "--last-modify-before" }, "Last modify time before fitler. e.g., 2022-08-02T16:20:21")
+        {
+            IsRequired = false,
+            Arity = ArgumentArity.ZeroOrOne
+        };
+    }
+
+    private Option<DateTime> CreateCreateAfterOption()
+    {
+        return new Option<DateTime>(new[] { "--create-after" }, "Create time after filter. e.g., 2022-07-01T10:20:21")
+        {
+            IsRequired = false,
+            Arity = ArgumentArity.ZeroOrOne
+        };
+    }
+
+    private Option<DateTime> CreateCreateBeforeOption()
+    {
+        return new Option<DateTime>(new[] { "--create-before" }, "Create time before fitler. e.g., 2022-12-02T16:20:21")
         {
             IsRequired = false,
             Arity = ArgumentArity.ZeroOrOne