+++ src/Option/OptionBase.cs
@@ -5,32 +5,14 @@ namespace Tur.Option;
 
 public abstract class OptionBase
 {
-    protected OptionBase(string outputDir, string[] includes, string[] excludes, bool enableVerbose,
-        string[] args, string cmd)
+    protected OptionBase(string[] args, string cmd)
     {
-        if (!string.IsNullOrEmpty(outputDir))
-        {
-            if (!Directory.Exists(outputDir))
-            {
-                throw new Exception($"Directory specified in -o/--output does not exist: {outputDir}");
-            }
-
-            OutputDir = Path.GetFullPath(outputDir);
-        }
-        else
-        {
-            OutputDir = Path.GetTempPath();
-        }
-
-        _ = Directory.CreateDirectory(OutputDir);
-        Includes = includes;
-        Excludes = excludes;
-        EnableVerbose = enableVerbose;
-        RawArgs = args;
         CmdName = cmd;
+        RawArgs = args;
+        OutputDir = Path.GetTempPath();
     }
 
-    public string OutputDir { get; set; }
+    public string OutputDir { get; set; } = Path.GetTempPath();
 
     public string[] Includes { get; set; }
 
@@ -40,9 +22,13 @@ public abstract class OptionBase
 
     public DateTime LastModifyAfter { get; set; }
 
+    public DateTime CreateBefore { get; set; }
+
+    public DateTime CreateAfter { get; set; }
+
     public bool EnableVerbose { get; set; }
 
-    public string[] RawArgs { get; set; }
+    public string[] RawArgs { get; }
 
     public string CmdName { get; }
 }