+++ src/Option/RmOption.cs
@@ -2,8 +2,7 @@
 
 public class RmOption : OptionBase
 {
-    public RmOption(string outputDir, string[] includes, string[] excludes, bool enableVerbose,
-        string[] args) : base(outputDir, includes, excludes, enableVerbose, args, "rm")
+    public RmOption(string[] args) : base(args, "rm")
     {
     }